Created Resources and decks to hold them in.
Added Resources and specific types.
Made decks and resource decks to sort and shuffle an arraylist of resources.
The player has a hand that will have multiple decks.

Made a few main for resource deck to see if shuffle & sort work correctly.

There is a demo for checking the board initialisation works too.